requirements.txt
opencv-python
pyserial
keyboard
cvzone
SpeechRecognition
PyAudio